pipeline/requirements.txt
feedparser>=6.0.11
google-genai>=1.0.0
python-dotenv>=1.0.0
